requirements.txt
requests
python-dotenv
python-dateutil
SQLAlchemy
